Reverted previous change - '=' is just as good as '==' in bash scripts, it seems.
The issue was that down in the DO_SHUTDOWN block, the reference to the
uninstall-os-vpx.sh script was referring to the "./scripts/..." path, which
is invalid, and returns an error.  Changed this reference to use $TOP_DIR like
the rest of the script.

Similarly, changed xenrc to use $TOP_DIR.